requirements.txt
djangorestframework
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.5
gunicorn==19.9.0
idna==2.8
lxml==4.3.1
nltk==3.4
numpy==1.16.1
praw==6.1.1
prawcore==1.0.1
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7.3
update-checker==0.16
urllib3==1.24.1
websocket-client==0.54.0
whitenoise==4.1.2
psycopg2
pandas==0.24.1
django-heroku
